updated index page
added deal banner to index. had to add a second index css page because the original was giving me problems. would not let me update any styles. had a logo to add but someone beat me to that. it would be nice to have a heads-up on things like that so my time will be wasted. still need to add more pictures and deal pictures to the deal element as well as quick links.